backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-socketio>=5.10.0
python-multipart>=0.0.6
pymongo>=4.6.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
nvidia-riva-client>=2.14.0
PyJWT>=2.8.0
httpx>=0.24.1
chromadb>=0.4.0,<0.6.0
openai>=1.0.0
smolagents
langchain
langchain-community
langchain-huggingface
langchain-chroma
sentence-transformers
tf-keras
pydub
logfire[fastapi,pymongo]
grpcio>=1.62.0
grpcio-tools>=1.62.0
